+++ src/plugins/piral-translate/src/create.ts
@@ -4,8 +4,14 @@ import type { PiralPlugin } from 'piral-core';
 import { createActions } from './actions';
 import { Localizer } from './localize';
 import { DefaultPicker } from './default';
-import { PiletLocaleApi, LocalizationMessages, Localizable, PiralSelectLanguageEvent, NestedLocalizationMessages } from './types';
 import { flattenTranslations } from './flatten-translations';
+import type {
+  PiletLocaleApi,
+  LocalizationMessages,
+  Localizable,
+  PiralSelectLanguageEvent,
+  AnyLocalizationMessages,
+} from './types';
 
 export interface TranslationFallback {
   (key: string, language: string): string;
@@ -23,7 +29,7 @@ export interface LocaleConfig {
    * Sets the default (global) localization messages.
    * @default {}
    */
-  messages?: LocalizationMessages | NestedLocalizationMessages;
+  messages?: AnyLocalizationMessages;
   /**
    * Sets the default language to use.
    */
@@ -72,18 +78,15 @@ export function createLocaleApi(localizer: Localizable = setupLocalizer()): Pira
     return (api) => {
       let localTranslations: LocalizationMessages = {};
 
+      const setTranslations = (messages: AnyLocalizationMessages) => {
+        localTranslations = flattenTranslations(messages);
+      };
+
       return {
         addTranslations(messages, isOverriding = true) {
-          const messagesToMerge = messages;
-
-          if (isOverriding) {
-            messagesToMerge.unshift(localizer.messages);
-          } else {
-            messagesToMerge.push(localizer.messages);
-          }
-
-          this.setTranslations(
-            deepmerge.all(messagesToMerge),
+          const current = localizer.messages;
+          setTranslations(
+            deepmerge.all<AnyLocalizationMessages>(isOverriding ? [current, ...messages] : [...messages, current]),
           );
         },
         getCurrentLanguage(cb?: (l: string) => void): any {
@@ -100,9 +103,7 @@ export function createLocaleApi(localizer: Localizable = setupLocalizer()): Pira
 
           return selected;
         },
-        setTranslations(messages) {
-          localTranslations = flattenTranslations(messages);
-        },
+        setTranslations,
         getTranslations() {
           return localTranslations;
         },