+++ src/plugins/piral-translate/src/flatten-translations.ts
@@ -1,23 +1,8 @@
-import { LocalizationMessages, NestedLocalizationMessages } from './types';
-
-export function flattenTranslations(messages: LocalizationMessages | NestedLocalizationMessages): LocalizationMessages {
-  return Object.fromEntries(
-    Object
-      .entries(messages)
-      .map(([ language, translations ]) => {
-        return [
-          language,
-          flat(translations)
-        ]
-      })
-  );
-}
+import type { LocalizationMessages, NestedLocalizationMessages } from './types';
 
 function flat(source: Record<string, unknown>): Record<string, string> {
   const target: Record<string, string> = {};
-
   flatten(source, target);
-
   return target;
 }
 
@@ -29,18 +14,18 @@ function flatten(source: any, target: Record<string, string>, prop = '') {
   }
 
   if (typeof source === 'object' && source !== null) {
-    Object
-      .keys(source)
-      .forEach((key) => {
-        flatten(
-          source[key],
-          target,
-          prop
-            ? `${prop}.${key}`
-            : key
-        );
-      });
+    Object.keys(source).forEach((key) => {
+      flatten(source[key], target, prop ? `${prop}.${key}` : key);
+    });
 
     return;
   }
+}
+
+export function flattenTranslations(messages: LocalizationMessages | NestedLocalizationMessages): LocalizationMessages {
+  return Object.fromEntries(
+    Object.entries(messages).map(([language, translations]) => {
+      return [language, flat(translations)];
+    }),
+  );
 }